pyproject.toml
[project]
name = "telephony-server"
version = "0.1.0"
description = "powerful bridge that connects telephony providers (Twilio, Vonage, Plivo, etc.) with real-time communication platforms (LiveKit, Jay.so, Pipecat, etc.). It enables seamless call routing, robust metrics collection, and observability features for enhanced telephony operations."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Navicstein Chinemerem", email = "navicsteinrotiv@gmail.com" },
]
dependencies = [
    "aiohttp>=3.11.10",
    "cuid2>=2.0.1",
    "fastapi>=0.115.8",
    "janus>=2.0.0",
    # "jay-ai",                                      # This package is having deps issues with other packages
    "loguru>=0.7.3",
    "openai>=1.58.1",
    "pipecat-ai[cartesia,openai,silero,deepgram]",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sentry-sdk>=2.20.0",
    "sounddevice>=0.5.1",
    "uvicorn>=0.34.0",
]
